+++ src/lib/components/json-schema/form/index.tsx
@@ -10,7 +10,7 @@ import {
   Widgets as DefaultWidgets,
 } from "@rjsf/chakra-ui";
 import type { FormProps } from "@rjsf/core";
-import type { RJSFSchema } from "@rjsf/utils";
+import type { RJSFSchema, RJSFValidationError } from "@rjsf/utils";
 import { createSchemaUtils } from "@rjsf/utils";
 import { customizeValidator } from "@rjsf/validator-ajv8";
 import isEqual from "lodash/isEqual";
@@ -99,7 +99,10 @@ export interface JsonSchemaFormProps
   initialFormData?: Record<string, unknown>;
   onSubmit?: (data: Record<string, unknown>) => void;
   /** Onchange callback is with BROKEN data */
-  onChange?: (data: Record<string, unknown>) => void;
+  onChange?: (
+    data: Record<string, unknown>,
+    errors: RJSFValidationError[]
+  ) => void;
   formContext?: Record<string, unknown>;
 }
 
@@ -136,13 +139,13 @@ export const JsonSchemaForm: FC<JsonSchemaFormProps> = ({
   };
 
   const onChange = useCallback(
-    (data: Record<string, unknown>) => {
+    (data: Record<string, unknown>, errors: RJSFValidationError[]) => {
       const values = { ...data };
       if (data) {
         fixOneOfKeysCallback(values);
         if (!isEqual(formData, values)) {
           setFormData(values);
-          propsOnChange?.(values);
+          propsOnChange?.(values, errors);
         }
       }
     },
@@ -184,9 +187,11 @@ export const JsonSchemaForm: FC<JsonSchemaFormProps> = ({
         ...templates,
       }}
       validator={v8Validator}
-      onChange={({ formData: values }) => {
+      liveValidate
+      showErrorList={false}
+      onChange={({ formData: values, errors }) => {
         // log.info(values)
-        onChange?.(values);
+        onChange?.(values, errors);
       }}
       onSubmit={({ formData: values }) => {
         // log.info(values)